requirements.txt
chex==0.1.86
tqdm
opencv-python
pillow
pandas
ipykernel
ipyfilechooser==0.6.0
ipython==8.12.3
ipywidgets==8.1.2
matplotlib==3.8.4
numpy==1.26.4
tensorflow_probability==0.23.0
plotly==6.2.0
scikit-learn==1.7.0
seaborn==0.13.2
termcolor==3.1.0
statsmodels==0.14.4
dtaidistance==2.3.13
pingouin==0.5.5
jax[cuda12]==0.5.3
genjax==0.10.3

# you also need FFMPEG